Move the RNN-T log-probability kernels out of parts/k2
The module depends on torch and Triton alone, and now has two callers rather than one,
so parts/k2 is no longer the right home for it. Relocate it beside the other Triton
kernels and repoint the four import sites.

Pure relocation: file contents and behaviour are unchanged.
